proprty list added
