+++ src/components/property/index.js
@@ -1,1 +1,1 @@
-export { default as PropertyList } from "./property-4/PropertyList";
+export { default as PropertyList } from "./PropertyList";
