--- src/pages/Portifolio.jsx
@@ -1,11 +1,0 @@
-import PortifolioList from "../components/portifolio/PortifolioList";
-
-const Portifolio = () => {
-  return (
-    <div className="pt-20 px-[3%] md:px-[6%]">
-      <PortifolioList />
-    </div>
-  );
-};
-
-export default Portifolio;
