--- src/pages/PortifolioTwo.jsx
@@ -1,11 +1,0 @@
-import PortifolioList from "../components/portifolio/PortifolioList";
-
-const PortifolioTwo = () => {
-  return (
-    <div className="pt-20 px-[3%] md:px-[6%]">
-      <PortifolioList flex />
-    </div>
-  );
-};
-
-export default PortifolioTwo;
