+++ src/pages/index.js
@@ -1,22 +1,4 @@
 export { default as Home } from "./Home";
-export { default as HomeTwo } from "./HomeTwo";
-export { default as About } from "./About";
-export { default as AboutTwo } from "./AboutTwo";
-export { default as Services } from "./Services";
-export { default as ServicesTwo } from "./ServicesTwo";
 export { default as Property } from "./Property";
-export { default as PropertyTwo } from "./PropertyTwo";
-export { default as PropertyThree } from "./PropertyThree";
-export { default as PropertyFour } from "./PropertyFour";
-export { default as PropertyFive } from "./PropertyFive";
-export { default as PropertySix } from "./PropertySix";
-export { default as Blog } from "./Blog";
-export { default as BlogTwo } from "./BlogTwo";
-export { default as BlogThree } from "./BlogThree";
-export { default as BlogFour } from "./BlogFour";
 export { default as Contact } from "./Contact";
-export { default as Portifolio } from "./Portifolio";
-export { default as PortifolioTwo } from "./PortifolioTwo";
 export { default as Team } from "./Team";
-export { default as Faqs } from "./Faqs";
-export { default as PageNotFound } from "./PageNotFound";
