+++ src/routes/PageRoutes.js
@@ -1,12 +1,14 @@
 import React from "react";
 import { Routes, Route } from "react-router";
 import Home from "../pages/Home";
+import { Property } from "../pages";
 
 const PageRoutes = () => {
   return (
     <>
       <Routes>
         <Route path="/" element={<Home />} />
+        <Route path="/property" element={<Property />} />
       </Routes>
     </>
   );
